Move yaml from shardingsphere-data-pipeline-api to shardingsphere-data-pipeline-core (#21361)
