+++ kernel/data-pipeline/core/src/main/java/org/apache/shardingsphere/data/pipeline/yaml/process/YamlPipelineReadConfiguration.java
@@ -15,37 +15,42 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.data.pipeline.api.config.process.yaml;
+package org.apache.shardingsphere.data.pipeline.yaml.process;
 
 import lombok.Data;
 import org.apache.shardingsphere.infra.util.yaml.YamlConfiguration;
 import org.apache.shardingsphere.infra.yaml.config.pojo.algorithm.YamlAlgorithmConfiguration;
 
 /**
- * YAML pipeline write configuration.
+ * YAML pipeline read configuration.
  */
 @Data
-public final class YamlPipelineWriteConfiguration implements YamlConfiguration {
+public final class YamlPipelineReadConfiguration implements YamlConfiguration {
     
     private static final Integer DEFAULT_WORKER_THREAD = 40;
     
     private static final Integer DEFAULT_BATCH_SIZE = 1000;
     
+    private static final Integer DEFAULT_SHARDING_SIZE = 1000_0000;
+    
     private Integer workerThread;
     
     private Integer batchSize;
+    
+    private Integer shardingSize;
     
     private YamlAlgorithmConfiguration rateLimiter;
     
     /**
      * Build with default value.
      *
-     * @return write configuration
+     * @return read configuration
      */
-    public static YamlPipelineWriteConfiguration buildWithDefaultValue() {
-        YamlPipelineWriteConfiguration result = new YamlPipelineWriteConfiguration();
+    public static YamlPipelineReadConfiguration buildWithDefaultValue() {
+        YamlPipelineReadConfiguration result = new YamlPipelineReadConfiguration();
         result.workerThread = DEFAULT_WORKER_THREAD;
         result.batchSize = DEFAULT_BATCH_SIZE;
+        result.shardingSize = DEFAULT_SHARDING_SIZE;
         return result;
     }
     
@@ -59,6 +64,9 @@ public final class YamlPipelineWriteConfiguration implements YamlConfiguration {
         if (null == batchSize) {
             batchSize = DEFAULT_BATCH_SIZE;
         }
+        if (null == shardingSize) {
+            shardingSize = DEFAULT_SHARDING_SIZE;
+        }
     }
     
     /**
@@ -66,7 +74,7 @@ public final class YamlPipelineWriteConfiguration implements YamlConfiguration {
      *
      * @param another another configuration
      */
-    public void copyNonNullFields(final YamlPipelineWriteConfiguration another) {
+    public void copyNonNullFields(final YamlPipelineReadConfiguration another) {
         if (null == another) {
             return;
         }
@@ -75,6 +83,9 @@ public final class YamlPipelineWriteConfiguration implements YamlConfiguration {
         }
         if (null != another.batchSize) {
             batchSize = another.batchSize;
+        }
+        if (null != another.shardingSize) {
+            shardingSize = another.shardingSize;
         }
         if (null != another.rateLimiter) {
             rateLimiter = another.rateLimiter;
@@ -93,6 +104,9 @@ public final class YamlPipelineWriteConfiguration implements YamlConfiguration {
                 break;
             case "BATCH_SIZE":
                 batchSize = null;
+                break;
+            case "SHARDING_SIZE":
+                shardingSize = null;
                 break;
             case "RATE_LIMITER":
                 rateLimiter = null;
